refactor: organize routes
