+++ wormhole-connect/src/components/RenderRows.tsx
@@ -3,7 +3,7 @@ import { makeStyles } from 'tss-react/mui';
 import Down from 'icons/Down';
 import { Collapse } from '@mui/material';
 import { joinClass } from 'utils/style';
-import { NestedRow } from 'utils/routes/types';
+import { NestedRow } from 'routes/types';
 
 const useStyles = makeStyles()((theme) => ({
   row: {
