+++ wormhole-connect/src/routes/bridge/index.ts
@@ -1,0 +1,2 @@
+export * from './baseRoute';
+export * from './bridge';
