+++ wormhole-connect/src/routes/cctpManual/cctpManual.ts
@@ -1,5 +1,4 @@
-import { BigNumber, utils, BytesLike, ethers } from 'ethers';
-import axios, { AxiosResponse } from 'axios';
+import { BigNumber, utils, ethers } from 'ethers';
 import {
   ChainId,
   ChainName,
@@ -10,20 +9,12 @@ import {
   MessageTransmitter__factory,
 } from '@wormhole-foundation/wormhole-connect-sdk';
 
-import {
-  CHAINS,
-  ROUTES,
-  TOKENS,
-  TOKENS_ARR,
-  isMainnet,
-  sdkConfig,
-} from 'config';
+import { CHAINS, ROUTES, TOKENS, TOKENS_ARR, sdkConfig } from 'config';
 import { TokenConfig, Route } from 'config/types';
 import {
   MAX_DECIMALS,
   getTokenById,
   getTokenDecimals,
-  sleep,
   toNormalizedDecimals,
   getDisplayName,
 } from 'utils';
@@ -39,96 +30,20 @@ import {
   UnsignedCCTPMessage,
   TransferInfoBaseParams,
   TransferDestInfoBaseParams,
-} from './types';
-import { BaseRoute } from './baseRoute';
-import { toDecimals } from '../balance';
-import { formatGasFee } from './utils';
+} from '../types';
+import { BaseRoute } from '../bridge/baseRoute';
+import { toDecimals } from '../../utils/balance';
+import { formatGasFee } from '../utils';
 import { getNativeVersionOfToken } from 'store/transferInput';
-
-export const CCTPTokenSymbol = 'USDC';
-export const CCTPManual_CHAINS: ChainName[] = [
-  'ethereum',
-  'avalanche',
-  'fuji',
-  'goerli',
-  'optimism',
-  'arbitrum',
-  'optimismgoerli',
-  'arbitrumgoerli',
-];
-export const CCTP_LOG_TokenMessenger_DepositForBurn =
-  '0x2fa9ca894982930190727e75500a97d8dc500233a5065e0f3126c48fbe0343c0';
-export const CCTP_LOG_MessageSent =
-  '0x8c5261668696ce22758910d05bab8f186d6eb247ceac2af2e82c7dc17669b036';
-
-export function getForeignUSDCAddress(chain: ChainName | ChainId) {
-  const usdcToken = TOKENS_ARR.find(
-    (t) =>
-      t.symbol === CCTPTokenSymbol &&
-      t.nativeChain === wh.toChainName(chain) &&
-      t.tokenId?.chain === wh.toChainName(chain),
-  );
-  if (!usdcToken) {
-    throw new Error('No foreign native USDC address');
-  }
-  return usdcToken.tokenId?.address;
-}
-
-const CIRCLE_ATTESTATION = isMainnet
-  ? 'https://iris-api.circle.com/attestations/'
-  : 'https://iris-api-sandbox.circle.com/attestations/';
-
-export async function getCircleAttestation(messageHash: BytesLike) {
-  while (true) {
-    // get the post
-    const response = await tryGetCircleAttestation(messageHash);
-
-    if (response) {
-      return response;
-    }
-
-    await sleep(6500);
-  }
-}
-
-export async function tryGetCircleAttestation(
-  messageHash: BytesLike,
-): Promise<string | undefined> {
-  return await axios
-    .get(`${CIRCLE_ATTESTATION}${messageHash}`)
-    .catch((reason) => {
-      return undefined;
-    })
-    .then(async (response: AxiosResponse | undefined) => {
-      if (
-        response &&
-        response.status === 200 &&
-        response.data.status === 'complete'
-      ) {
-        return response.data.attestation as string;
-      }
-
-      return undefined;
-    });
-}
-
-export function getChainNameCCTP(domain: number): ChainName {
-  switch (domain) {
-    case 0:
-      return isMainnet ? 'ethereum' : 'goerli';
-    case 1:
-      return isMainnet ? 'avalanche' : 'fuji';
-    case 2:
-      return isMainnet ? 'optimism' : 'optimismgoerli';
-    case 3:
-      return isMainnet ? 'arbitrum' : 'arbitrumgoerli';
-  }
-  throw new Error('Invalid CCTP domain');
-}
-
-export function getNonce(message: string): number {
-  return BigNumber.from('0x' + message.substring(26, 42)).toNumber();
-}
+import {
+  CCTPManual_CHAINS,
+  CCTPTokenSymbol,
+  CCTP_LOG_MessageSent,
+  CCTP_LOG_TokenMessenger_DepositForBurn,
+  getChainNameCCTP,
+  getNonce,
+  tryGetCircleAttestation,
+} from './utils';
 
 export class CCTPManualRoute extends BaseRoute {
   readonly NATIVE_GAS_DROPOFF_SUPPORTED: boolean = false;
@@ -674,23 +589,6 @@ export class CCTPManualRoute extends BaseRoute {
     );
     const result = await contract.usedNonces(hash);
     return result.toString() === '1';
-  }
-
-  async nativeTokenAmount(
-    destChain: ChainName | ChainId,
-    token: TokenId,
-    amount: BigNumber,
-    walletAddress: string,
-  ): Promise<BigNumber> {
-    throw new Error('Not supported');
-  }
-
-  async maxSwapAmount(
-    destChain: ChainName | ChainId,
-    token: TokenId,
-    walletAddress: string,
-  ): Promise<BigNumber> {
-    throw new Error('Not supported');
   }
 
   async tryFetchRedeemTx(
