+++ wormhole-connect/src/routes/cctpManual/index.ts
@@ -1,0 +1,2 @@
+export * from './cctpManual';
+export * from './utils';
