+++ wormhole-connect/src/routes/cctpRelay/index.ts
@@ -1,0 +1,1 @@
+export * from './cctpRelay';
