+++ wormhole-connect/src/routes/cosmosGateway/index.ts
@@ -1,0 +1,3 @@
+export * from './cosmosGateway';
+export * from './utils';
+export * from './types';
