+++ wormhole-connect/src/routes/cosmosGateway/utils/contracts.ts
@@ -1,0 +1,41 @@
+import { ChainId, ChainName } from '@wormhole-foundation/wormhole-connect-sdk';
+import { CHAIN_ID_WORMCHAIN, isCosmWasmChain } from '@certusone/wormhole-sdk';
+import { CHAINS } from 'config';
+import { wh } from 'utils/sdk';
+import { getCosmWasmClient, getQueryClient } from './client';
+import { IBC_PORT } from './consts';
+
+export function getTranslatorAddress(): string {
+  const addr = CHAINS['wormchain']?.contracts.ibcShimContract;
+  if (!addr) throw new Error('IBC Shim contract not configured');
+  return addr;
+}
+
+export async function getIbcDestinationChannel(
+  chain: ChainId | ChainName,
+): Promise<string> {
+  const sourceChannel = await getIbcSourceChannel(chain);
+  const queryClient = await getQueryClient(CHAIN_ID_WORMCHAIN);
+  const conn = await queryClient.ibc.channel.channel(IBC_PORT, sourceChannel);
+
+  const destChannel = conn.channel?.counterparty?.channelId;
+  if (!destChannel) {
+    throw new Error(`No destination channel found on chain ${chain}`);
+  }
+
+  return destChannel;
+}
+
+export async function getIbcSourceChannel(
+  chain: ChainId | ChainName,
+): Promise<string> {
+  const id = wh.toChainId(chain);
+  if (!isCosmWasmChain(id)) throw new Error('Chain is not cosmos chain');
+  const client = await getCosmWasmClient(CHAIN_ID_WORMCHAIN);
+  const { channel } = await client.queryContractSmart(getTranslatorAddress(), {
+    ibc_channel: {
+      chain_id: id,
+    },
+  });
+  return channel;
+}
