+++ wormhole-connect/src/routes/cosmosGateway/utils/index.ts
@@ -1,0 +1,8 @@
+export * from './client';
+export * from './consts';
+export * from './events';
+export * from './fromCosmos';
+export * from './toCosmos';
+export * from './transaction';
+export * from './getMessage';
+export * from './contracts';
