+++ wormhole-connect/src/routes/index.ts
@@ -1,4 +1,4 @@
-export * from './routeAbstract';
+export * from './abstracts';
 export * from './bridge';
 export * from './relay';
 export * from './hashflow';
