+++ wormhole-connect/src/routes/relay/index.ts
@@ -1,0 +1,2 @@
+export * from './relay';
+export * from './types';
