+++ wormhole-connect/src/routes/relay/types.ts
@@ -1,0 +1,13 @@
+import { ParsedMessage, ParsedRelayerMessage } from 'utils/sdk';
+
+export type RelayOptions = {
+  relayerFee?: number;
+  toNativeToken?: number;
+  receiveNativeAmt: number;
+};
+
+export interface TransferDestInfoParams {
+  txData: ParsedMessage | ParsedRelayerMessage;
+  receiveTx?: string;
+  transferComplete?: boolean;
+}
