+++ wormhole-connect/src/utils/events.ts
@@ -13,9 +13,9 @@ import { fromNormalizedDecimals } from '.';
 import { CHAINS } from 'config';
 import { fetchGlobalTx, getEmitterAndSequence } from './vaa';
 import { isEvmChain } from 'utils/sdk';
-import RouteOperator from './routes/operator';
+import RouteOperator from '../routes/operator';
 import { Route } from 'config/types';
-import { SignedMessage } from './routes';
+import { SignedMessage } from '../routes';
 
 export const fetchRedeemTx = async (
   route: Route,
