+++ wormhole-connect/src/views/Redeem/Redeem.tsx
@@ -10,8 +10,8 @@ import {
 } from 'store/redeem';
 import { sleep } from 'utils';
 import { fetchIsVAAEnqueued } from 'utils/vaa';
-import { SignedMessage } from 'utils/routes';
-import RouteOperator from 'utils/routes/operator';
+import { SignedMessage } from 'routes';
+import RouteOperator from 'routes/operator';
 import { ParsedMessage, ParsedRelayerMessage } from 'utils/sdk';
 
 import PageHeader from 'components/PageHeader';
