+++ wormhole-connect/src/views/Redeem/SendFrom.tsx
@@ -4,8 +4,8 @@ import { useSelector } from 'react-redux';
 import InputContainer from 'components/InputContainer';
 import { RenderRows } from 'components/RenderRows';
 import { RootState } from 'store';
-import { TransferDisplayData } from 'utils/routes';
-import RouteOperator from 'utils/routes/operator';
+import { TransferDisplayData } from 'routes';
+import RouteOperator from 'routes/operator';
 
 import Confirmations from './Confirmations';
 import Header from './Header';
