+++ wormhole-connect/src/views/TxSearch.tsx
@@ -6,7 +6,7 @@ import { ChainName } from '@wormhole-foundation/wormhole-connect-sdk';
 
 import { CHAINS_ARR } from 'config';
 import { isValidTxId } from 'utils';
-import RouteOperator from 'utils/routes/operator';
+import RouteOperator from 'routes/operator';
 import { setTxDetails, setRoute as setRedeemRoute } from 'store/redeem';
 import { setRoute as setAppRoute } from 'store/router';
 import PageHeader from 'components/PageHeader';
